v.12

+++ MoviesAIModel/myapp/AIModel/Models/UserPreference.py
@@ -5,7 +5,7 @@ import numpy as np
 
 
 class UserPreference:
-    def __init__(self, id: str, user_id: int, preference: Optional[List[int]], rated_film_ids: Optional[List[UUID]]):
+    def __init__(self, id: str, user_id: int, preference: Optional[List[float]], rated_film_ids: Optional[List[UUID]]):
         self.id = id
         self.user_id = user_id
         self.preference = preference
